+++ sites/funmooc/src/backend/funmooc/urls.py
@@ -11,6 +11,7 @@ from django.views.generic import TemplateView
 from django.views.static import serve
 
 from cms.sitemaps import CMSSitemap
+from richie.apps.courses.urls import urlpatterns as courses_urlpatterns
 from richie.apps.search.urls import urlpatterns as search_urlpatterns
 
 # For now, we use URLPathVersioning to be consistent with fonzie. Fonzie uses it
@@ -25,7 +26,7 @@ urlpatterns = [
     path(r"sitemap.xml", sitemap, {"sitemaps": {"cmspages": CMSSitemap}}),
     re_path(
         r"api/{}/".format(API_PREFIX),
-        include([*search_urlpatterns]),
+        include([*courses_urlpatterns, *search_urlpatterns]),
     ),
     path(r"", include("filer.server.urls")),
 ]